+++ src/main/java/ca/uwaterloo/drinkmasterapi/feature/order/service/OrderServiceImpl.java
@@ -2,14 +2,14 @@ package ca.uwaterloo.drinkmasterapi.feature.order.service;
 
 import ca.uwaterloo.drinkmasterapi.common.ResourceNotFoundException;
 import ca.uwaterloo.drinkmasterapi.dao.Drink;
-import ca.uwaterloo.drinkmasterapi.feature.drink.repository.DrinkRepository;
-import ca.uwaterloo.drinkmasterapi.feature.mqtt.repository.MachineRepository;
+import ca.uwaterloo.drinkmasterapi.repository.DrinkRepository;
+import ca.uwaterloo.drinkmasterapi.repository.MachineRepository;
 import ca.uwaterloo.drinkmasterapi.dao.Order;
 import ca.uwaterloo.drinkmasterapi.feature.order.model.OrderRequestDTO;
 import ca.uwaterloo.drinkmasterapi.feature.order.model.OrderResponseDTO;
 import ca.uwaterloo.drinkmasterapi.common.OrderStatusEnum;
-import ca.uwaterloo.drinkmasterapi.feature.order.repository.OrderRepository;
-import ca.uwaterloo.drinkmasterapi.feature.user.repository.UserRepository;
+import ca.uwaterloo.drinkmasterapi.repository.OrderRepository;
+import ca.uwaterloo.drinkmasterapi.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
